add functions to sign the new TransactionMessage type (#2387)
* add a new ITransactionWithSignatures and IFullySignedTransaction type

* remove the OrderedMap class and use native objects

* rename new transaction/fully signed transaction types

* refactor to sign a NewTransaction and not a TransactionMessage

* refactor to always compute a new signature and replace if changed (or null)

* throw an error if partialSign is called with a keypair that is not expected to sign the transaction

* refactor new-signatures unit tests to use await expect

+++ packages/transactions/src/signatures.ts
@@ -46,10 +46,7 @@ export async function partiallySignTransaction<TTransaction extends CompilableTr
     const wireMessageBytes = getCompiledMessageEncoder().encode(compiledMessage);
     const publicKeySignaturePairs = await Promise.all(
         keyPairs.map(keyPair =>
-            Promise.all([
-                getAddressFromPublicKey(keyPair.publicKey),
-                signBytes(keyPair.privateKey, wireMessageBytes as Uint8Array),
-            ]),
+            Promise.all([getAddressFromPublicKey(keyPair.publicKey), signBytes(keyPair.privateKey, wireMessageBytes)]),
         ),
     );
     for (const [signerPublicKey, signature] of publicKeySignaturePairs) {
